+++ PearlLib/PearlLib/pearl_loss.h
@@ -11,15 +11,15 @@ typedef enum {
 
 typedef struct {
     pearl_loss_type type;
-    double (*calculate)(double, double);
-    double (*calculate_derivative)(double, double);
+    float (*calculate)(float, float);
+    float (*calculate_derivative)(float, float);
 } pearl_loss;
 
 PEARL_API pearl_loss pearl_loss_create(pearl_loss_type type);
-double pearl_loss_cost(pearl_loss loss, const pearl_tensor *output, const pearl_tensor *output_prediction);
-double pearl_loss_binary_cross_entropy_func(double out, double pred);
-double pearl_loss_mean_squared_error_func(double out, double pred);
-double pearl_loss_binary_cross_entropy_func_derivative(double out, double pred);
-double pearl_loss_mean_squared_error_func_derivative(double out, double pred);
+float pearl_loss_cost(pearl_loss loss, const pearl_tensor *output, const pearl_tensor *output_prediction);
+float pearl_loss_binary_cross_entropy_func(float out, float pred);
+float pearl_loss_mean_squared_error_func(float out, float pred);
+float pearl_loss_binary_cross_entropy_func_derivative(float out, float pred);
+float pearl_loss_mean_squared_error_func_derivative(float out, float pred);
 
 #endif // PEARL_LOSS_H
